# Map performance - use pattern match on size to reduce recursive function calls