@HuwCampbell 2024-11-14
The version you posted is about 10 to 15% slower than what's there already. See [this commit](https://github.com/haskell/containers/commit/29d3fbcc67ea4eebc7b381ea5c59673ef1e10dc9) which adds the change it to `foldMap`.

It's worth asking why it's slower – it seems that pattern matching on the strict packed integer should be very fast, as there's no pointer chasing or indirection involved, pattern matching on the tips does require some indirection.

From the comment in the link above, eliding 2/3 of the recursive calls to Tip is a big deal, because there are a _lot_ of Tips, effectively about the same number as the number of items in the map.

Maybe we should add a GHC style note about this optimisation.